done permission

+++ src/main/java/com/cg/coffeemanagement/controller/api/StaffApi.java
@@ -1,13 +1,17 @@
 package com.cg.coffeemanagement.controller.api;
 
+import com.cg.coffeemanagement.Static.Principal;
 import com.cg.coffeemanagement.exception.DataInputException;
 import com.cg.coffeemanagement.model.Position;
 import com.cg.coffeemanagement.model.Staff;
+import com.cg.coffeemanagement.model.User;
 import com.cg.coffeemanagement.model.dto.StaffDto;
 import com.cg.coffeemanagement.services.Positions.IPositionServices;
 import com.cg.coffeemanagement.services.Staffs.IStaffServices;
+import com.cg.coffeemanagement.services.Users.IUserService;
 import com.cg.coffeemanagement.utils.AppUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
@@ -27,6 +31,8 @@ public class StaffApi {
     private IPositionServices positionServices;
     @Autowired
     private AppUtil appUtil;
+    @Autowired
+    private IUserService userService;
 
     @GetMapping
     public ResponseEntity<?> getStaffActive(){
@@ -86,6 +92,10 @@ public class StaffApi {
     public ResponseEntity<?> doDelete(@PathVariable Long id, @Validated @RequestBody StaffDto staffDto) {
         Optional<Staff> opStaff = staffServices.findById(id);
         if (opStaff.isPresent()) {
+            User presentUser = userService.getByUsername(Principal.getPrincipal()).get();
+            if (presentUser.getStaff().getPosition().getPermission().getPermissionAccess() >= opStaff.get().getPosition().getPermission().getPermissionAccess()){
+                throw new DataInputException("Không đủ quyền để thực hiện thao tác này");
+            }
             staffServices.deleteStaff(id);
             return new ResponseEntity<>(HttpStatus.OK);
         } else {
@@ -101,6 +111,10 @@ public class StaffApi {
         }
         Optional<Staff> opStaff = staffServices.findById(id);
         if (opStaff.isPresent()) {
+            User presentUser = userService.getByUsername(Principal.getPrincipal()).get();
+            if (presentUser.getStaff().getPosition().getPermission().getPermissionAccess() >= opStaff.get().getPosition().getPermission().getPermissionAccess()){
+                throw new DataInputException("Không đủ quyền để thực hiện thao tác này");
+            }
             Staff staff = opStaff.get();
             staff.setName(staffDto.getName());
             staff.setAddress(staffDto.getAddress());
